to #246 feat: add object_storage_client.go.
